+++ cohere/utils.py
@@ -3,7 +3,7 @@ import json
 try:  # numpy is optional, but support json encoding if the user has it
     import numpy as np
 
-    class NumpyAwareJsonEncoder(json.JSONEncoder):
+    class CohereJsonEncoder(json.JSONEncoder):
         """Handles numpy datatypes and such in json encoding"""
 
         def default(self, obj):
@@ -20,8 +20,8 @@ try:  # numpy is optional, but support json encoding if the user has it
 
 except:
 
-    class NumpyAwareJsonEncoder(json.JSONEncoder):
-        """Handles numpy datatypes and such in json encoding"""
+    class CohereJsonEncoder(json.JSONEncoder):
+        """numpy is missing, so we can't handle these (and don't expect them)"""
 
         def default(self, obj):
             if isinstance(obj, set):
@@ -31,4 +31,4 @@ except:
 
 
 def np_json_dumps(data, **kwargs):
-    return json.dumps(data, cls=NumpyAwareJsonEncoder, **kwargs)
+    return json.dumps(data, cls=CohereJsonEncoder, **kwargs)