+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.0.2
+- [#183](https://github.com/cohere-ai/cohere-python/pull/183)
+  - Better error messages for synchronous client
+
+## 4.0.1
+- [#181](https://github.com/cohere-ai/cohere-python/pull/181)
+  - Allow Python >=3.11
+
 ## 4.0
 
 - [#160](https://github.com/cohere-ai/cohere-python/pull/160)
@@ -13,8 +21,11 @@
   - Use Poetry, and format using black and isort, include pre-commit hooks.
   - Removed ability for user to choose API version. This SDK version defaults to v1.
   - Added 'meta' fields to response objects with API version
-  - Updated co.Chat to use persona_name, persona_prompt, user_name
-  
+- [#179](https://github.com/cohere-ai/cohere-python/pull/179)
+  - Add support for co.Chat parameters: `temperature`, `max_tokens`, `persona_name`, `persona_prompt`
+  - Remove support for co.Chat parameters: `persona`, `preamble_override`
+  - Updates the co.Chat `user_name` parameter
+
 ## 3.10.0
 - [#176](https://github.com/cohere-ai/cohere-python/pull/176)
   - Add failure reason to clustering jobs